refactors logging api samples
